refactor: move Creator download handlers to a shared helper
Per review, keep Form.vue isolated: it now only re-emits
download-zip-file, download-input-file and show-update-success-message,
and problem/new.ts handles them. The download/notification logic is
extracted into problem/creator/downloadHandlers.ts and shared by both
new.ts and creator.ts, removing the duplication.
